feat: Implementado método que gera um relatório financeiro com base e…

+++ CleanArchitecture.OCP/Domain/DomainServices/FinancialReport.cs
@@ -1,8 +1,8 @@
 ﻿using CleanArchitecture.OCP.Domain.Entities;
 
 namespace CleanArchitecture.OCP.Domain.DomainServices;
-public class FinancialReport(IEnumerable<FinancialTransaction> financialTransactions, int total)
+public class FinancialReport(IEnumerable<FinancialTransaction> financialTransactions, decimal total)
 {
     public IEnumerable<FinancialTransaction> FinancialTransactions { get; } = financialTransactions;
-    public int Total { get; } = total;
+    public decimal Total { get; } = total;
 }
